# backend/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Costs manager API",
    "description": "RESTful API service for costs management"
  },
  "host": "ec2-54-221-33-134.compute-1.amazonaws.com:3000",
  "paths": {
    "/api/users": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "first_name": {
                  "type": "string",
                  "required": "true",
                  "example": "Israel"
                },
                "last_name": {
                  "type": "string",
                  "required": "true",
                  "example": "Israeli"
                },
                "martial_status": {
                  "type": "string",
                  "required": "true",
                  "example": "Single"
                },
                "birth_date": {
                  "type": "string",
                  "required": "true",
                  "example": "1990-06-01"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User created."
          },
          "400": {
            "description": "Bad request, missing or invalid parameters."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/api/costs": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "user_id": {
                  "type": "string",
                  "required": "true",
                  "example": "<user_id_returned_from_create_user>"
                },
                "date": {
                  "type": "string",
                  "example": "2022-05-01T16:00:00"
                },
                "price": {
                  "type": "integer",
                  "required": "true",
                  "example": "200"
                },
                "category": {
                  "type": "string",
                  "required": "true",
                  "example": "Clothing"
                },
                "description": {
                  "type": "string",
                  "required": "true",
                  "example": "T-Shirt"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cost item created."
          },
          "400": {
            "description": "Bad request, missing or invalid parameters."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/api/report": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "user_id",
            "required": "true",
            "in": "query",
            "type": "string"
          },
          {
            "name": "year",
            "required": "true",
            "in": "query",
            "type": "string"
          },
          {
            "name": "month",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Cost item created."
          },
          "400": {
            "description": "Bad request, missing or invalid parameters."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    }
  }
}